--- app/flaskr/__init__.py
@@ -1,25 +1,0 @@
-from flask import Flask
-
-# def create_app():
-#     app = Flask(__name__)
-
-#     @app.route('/')
-#     def hello_world():
-#         return 'Hello, World!'
-
-#     from . import db
-#     db.init_app(app)
-
-#     return app
-
-#     if __name__ == '__main__':
-#     app.run(host='0.0.0.0')
-
-def create_app():
-    app = ...
-    # existing code omitted
-
-    from . import db
-    db.init_app(app)
-
-    return app
