+++ app/templates/layout.html
@@ -1,0 +1,3 @@
+{% block heading %}{% endblock %}
+
+{% block content %}{% endblock %}
