+++ app/templates/tutor.html
@@ -1,0 +1,32 @@
+{% extends 'layout.html' %} {% block heading %}
+
+Tutor Information: {{tutors.name}}{% endblock %} {% block content %}
+Name: {{Tutors.user_name}}
+ID: {{Tutors.user_id}}
+
+School: {{Tutors.school}}
+Age: {{Tutors.age}}
+Grade: {{Tutors.grade}}
+
+Bio: {{tutor.bio}}
+Rating: {{tutor.rating}}
+
+Location: {{Tutors.location}}
+Phone Number: {{Tutors.phone_number}}
+Email: {{Tutors.email}}
+
+Venmo: {{Tutors.venmo}}
+Hourly Rate: {{Tutors.hourly_rate}}
+ 
+
+Can Tutor In:
+
+{% for CanTutor in user_id.CanTutorIn %}
+{{CanTutor.class_id}}
+{% endfor %}
+
+Tutors in:
+
+{% for Tutors in user_id.TutorsIn %}
+{{Tutors.zoom_link}}
+{% endfor %}
